Rename and refactor Haskell transform tests
