--- tools/a2mochi/x/hs/convert.go
@@ -1,775 +1,0 @@
-//go:build slow
-
-package hs
-
-import (
-	"fmt"
-	"regexp"
-	"strings"
-
-	"mochi/ast"
-	"mochi/parser"
-	"mochi/transpiler/meta"
-)
-
-var traceNames map[string]bool
-
-func replaceOutsideQuotes(s, name, repl string) string {
-	parts := strings.Split(s, "\"")
-	re := regexp.MustCompile(`\b` + name + `\b`)
-	for i := 0; i < len(parts); i += 2 {
-		parts[i] = re.ReplaceAllString(parts[i], repl)
-	}
-	return strings.Join(parts, "\"")
-}
-
-// Field describes a record field in a data declaration.
-type Field struct {
-	Name string `json:"name"`
-	Type string `json:"type"`
-}
-
-// Item represents a minimal Haskell AST item parsed from source code.
-type Item struct {
-	Kind       string   `json:"kind"`
-	Name       string   `json:"name,omitempty"`
-	Params     []string `json:"params,omitempty"`
-	Body       string   `json:"body,omitempty"`
-	Type       string   `json:"type,omitempty"`
-	Fields     []Field  `json:"fields,omitempty"`
-	Collection string   `json:"collection,omitempty"`
-	Start      string   `json:"start,omitempty"`
-	End        string   `json:"end,omitempty"`
-	Line       int      `json:"line"`
-	Doc        string   `json:"doc,omitempty"`
-	Derives    []string `json:"derives,omitempty"`
-}
-
-// Parse parses a very small subset of Haskell source and returns a slice
-// of Items describing top level declarations. Unsupported lines produce an
-// error that includes the line number and snippet.
-func Parse(src string) ([]Item, error) {
-	lines := strings.Split(src, "\n")
-	var items []Item
-	var parseErr error
-	sigs := make(map[string]string)
-	for i := 0; i < len(lines); i++ {
-		line := lines[i]
-		l := strings.TrimSpace(line)
-		if strings.HasPrefix(line, " ") || strings.HasPrefix(line, "\t") {
-			continue
-		}
-		if l == "" || strings.HasPrefix(l, "--") || strings.HasPrefix(l, "module ") || strings.HasPrefix(l, "import ") || strings.HasPrefix(l, "{-") {
-			continue
-		}
-		if parts := strings.SplitN(l, "::", 2); len(parts) == 2 {
-			name := strings.TrimSpace(parts[0])
-			sigs[name] = strings.TrimSpace(parts[1])
-			continue
-		}
-		if strings.HasPrefix(l, "main =") {
-			body := strings.TrimSpace(strings.TrimPrefix(l, "main ="))
-			if body == "do" {
-				j := i + 1
-				for j < len(lines) {
-					ln := lines[j]
-					if strings.TrimSpace(ln) == "" {
-						j++
-						continue
-					}
-					if !strings.HasPrefix(ln, " ") && !strings.HasPrefix(ln, "\t") {
-						break
-					}
-					t := strings.TrimSpace(ln)
-					if strings.HasPrefix(t, "print") {
-						arg := strings.TrimSpace(strings.TrimPrefix(t, "print"))
-						if strings.HasPrefix(arg, "(") && strings.HasSuffix(arg, ")") {
-							arg = strings.TrimSuffix(strings.TrimPrefix(arg, "("), ")")
-						}
-						items = append(items, Item{Kind: "print", Body: arg, Line: j + 1})
-					} else if strings.HasPrefix(t, "putStrLn") {
-						arg := strings.TrimSpace(strings.TrimPrefix(t, "putStrLn"))
-						if strings.HasPrefix(arg, "(") && strings.HasSuffix(arg, ")") {
-							arg = strings.TrimSuffix(strings.TrimPrefix(arg, "("), ")")
-						}
-						if strings.HasPrefix(arg, "show") {
-							arg = "str(" + strings.TrimSpace(strings.TrimPrefix(arg, "show ")) + ")"
-						}
-						items = append(items, Item{Kind: "print", Body: arg, Line: j + 1})
-					} else if strings.HasPrefix(t, "_json") {
-						arg := strings.TrimSpace(strings.TrimPrefix(t, "_json"))
-						if strings.HasPrefix(arg, "(") && strings.HasSuffix(arg, ")") {
-							arg = strings.TrimSuffix(strings.TrimPrefix(arg, "("), ")")
-						}
-						items = append(items, Item{Kind: "json", Body: arg, Line: j + 1})
-					} else if strings.HasPrefix(t, "mapM_") {
-						buf := t
-						k := j
-						openParen := strings.Count(buf, "(") - strings.Count(buf, ")")
-						openBrack := strings.Count(buf, "[") - strings.Count(buf, "]")
-						for (openParen > 0 || openBrack > 0) && k+1 < len(lines) {
-							k++
-							next := strings.TrimSpace(lines[k])
-							buf += " " + next
-							openParen = strings.Count(buf, "(") - strings.Count(buf, ")")
-							openBrack = strings.Count(buf, "[") - strings.Count(buf, "]")
-						}
-						if it, ok := parseMapM(buf, k+1); ok {
-							items = append(items, it)
-							j = k
-						}
-					}
-					j++
-				}
-				i = j - 1
-				continue
-			}
-			if strings.HasPrefix(body, "putStrLn") {
-				arg := strings.TrimSpace(strings.TrimPrefix(body, "putStrLn"))
-				if strings.HasPrefix(arg, "(") && strings.HasSuffix(arg, ")") {
-					arg = strings.TrimSuffix(strings.TrimPrefix(arg, "("), ")")
-				}
-				items = append(items, Item{Kind: "print", Body: arg, Line: i + 1})
-			}
-			continue
-		}
-		if strings.HasPrefix(l, "data ") && strings.Contains(l, "{") {
-			parts := strings.Fields(strings.TrimSpace(strings.TrimPrefix(l, "data ")))
-			if len(parts) > 0 {
-				name := parts[0]
-				j := i
-				buf := l
-				for j+1 < len(lines) && !strings.Contains(lines[j], "}") {
-					j++
-					buf += "\n" + strings.TrimSpace(lines[j])
-				}
-				if end := strings.Index(buf, "}"); end != -1 {
-					start := strings.Index(buf, "{")
-					fields := strings.Split(buf[start+1:end], ",")
-					var fs []Field
-					for _, f := range fields {
-						f = strings.TrimSpace(f)
-						if f == "" {
-							continue
-						}
-						parts := strings.Fields(f)
-						if len(parts) >= 3 && parts[1] == "::" {
-							fs = append(fs, Field{Name: parts[0], Type: parts[2]})
-						}
-					}
-					var derives []string
-					rest := strings.TrimSpace(buf[end+1:])
-					if rest == "" && j+1 < len(lines) {
-						next := strings.TrimSpace(lines[j+1])
-						if strings.HasPrefix(next, "deriving") {
-							rest = next
-							j++
-						}
-					}
-					if strings.HasPrefix(rest, "deriving") {
-						list := rest[len("deriving"):]
-						list = strings.TrimSpace(strings.Trim(list, "()"))
-						parts := strings.Split(list, ",")
-						for _, p := range parts {
-							p = strings.TrimSpace(p)
-							if p != "" {
-								derives = append(derives, p)
-							}
-						}
-					}
-					items = append(items, Item{Kind: "struct", Name: name, Fields: fs, Line: i + 1, Derives: derives})
-					i = j
-					continue
-				}
-			}
-		}
-		if strings.HasPrefix(l, "data ") || strings.HasPrefix(l, "newtype ") || strings.HasPrefix(l, "type ") || strings.HasPrefix(l, "instance ") {
-			continue
-		}
-		if strings.HasPrefix(l, "|") || strings.HasPrefix(l, "case ") || strings.HasPrefix(l, "of ") || l == "where" {
-			continue
-		}
-		if parts := strings.SplitN(l, "=", 2); len(parts) == 2 {
-			left := strings.Fields(strings.TrimSpace(parts[0]))
-			if len(left) == 0 {
-				continue
-			}
-			name := left[0]
-			switch name {
-			case "forLoop", "whileLoop", "avg", "_group_by", "_indexString", "_input", "_now", "_json", "_readInput", "_writeOutput", "_split", "_parseCSV", "_load", "_save",
-				"_asInt", "_asDouble", "_asString", "_asBool", "_parseJSON", "_valueToMap", "_valueToString", "_mapToValue":
-				continue
-			}
-			params := left[1:]
-			body := strings.TrimSpace(parts[1])
-			if strings.HasPrefix(body, "do") {
-				continue
-			}
-			kind := "func"
-			if len(params) == 0 {
-				kind = "var"
-			}
-			it := Item{Kind: kind, Name: name, Params: params, Body: body, Line: i + 1}
-			if t, ok := sigs[name]; ok {
-				it.Type = t
-			}
-			items = append(items, it)
-			continue
-		}
-		if parseErr == nil {
-			start := i - 1
-			if start < 0 {
-				start = 0
-			}
-			end := i + 1
-			if end >= len(lines) {
-				end = len(lines) - 1
-			}
-			var snippet []string
-			for j := start; j <= end; j++ {
-				prefix := "   "
-				if j == i {
-					prefix = "-> "
-				}
-				snippet = append(snippet, fmt.Sprintf("%s%3d | %s", prefix, j+1, lines[j]))
-			}
-			parseErr = fmt.Errorf("line %d: unsupported syntax: %s\n%s", i+1, strings.TrimSpace(lines[i]), strings.Join(snippet, "\n"))
-		}
-	}
-	return items, parseErr
-}
-
-// parseMapM parses a line using mapM_ loop syntax and returns an Item describing a for loop.
-func parseMapM(line string, lineNum int) (Item, bool) {
-	l := strings.TrimSpace(strings.TrimPrefix(line, "mapM_"))
-	if !strings.HasPrefix(l, "(\\") {
-		return Item{}, false
-	}
-	idx := strings.Index(l, "->")
-	if idx == -1 {
-		return Item{}, false
-	}
-	varName := strings.TrimSpace(l[2:idx])
-	rest := strings.TrimSpace(l[idx+2:])
-	depth := 1
-	close := -1
-	for i := 0; i < len(rest); i++ {
-		switch rest[i] {
-		case '(':
-			depth++
-		case ')':
-			depth--
-			if depth == 0 {
-				close = i
-				i = len(rest)
-				break
-			}
-		}
-	}
-	if close == -1 {
-		return Item{}, false
-	}
-	body := strings.TrimSpace(rest[:close])
-	if strings.HasPrefix(body, "do") {
-		body = strings.TrimSpace(strings.TrimPrefix(body, "do"))
-	}
-	src := strings.TrimSpace(rest[close+1:])
-	if strings.HasPrefix(src, "(") && strings.HasSuffix(src, ")") {
-		src = strings.TrimSuffix(strings.TrimPrefix(src, "("), ")")
-	}
-	if strings.HasPrefix(src, "[") && strings.HasSuffix(src, "]") {
-		s := strings.TrimSuffix(strings.TrimPrefix(src, "["), "]")
-		if parts := strings.SplitN(s, "..", 2); len(parts) == 2 {
-			start := strings.TrimSpace(parts[0])
-			end := strings.TrimSpace(parts[1])
-			end = strings.TrimSuffix(end, "- 1")
-			end = strings.TrimSpace(end)
-			end = "(" + end + " + 1)"
-			return Item{Kind: "for", Name: varName, Start: start, End: end, Body: convertBody(body), Line: lineNum}, true
-		}
-	}
-	if strings.HasPrefix(src, "Map.keys ") {
-		src = strings.TrimPrefix(src, "Map.keys ")
-	}
-	return Item{Kind: "for", Name: varName, Collection: src, Body: convertBody(body), Line: lineNum}, true
-}
-
-func convertBody(body string) string {
-	if strings.HasPrefix(body, "print") {
-		arg := strings.TrimSpace(strings.TrimPrefix(body, "print"))
-		arg = strings.Trim(arg, "()")
-		return "print(" + convertExpr(arg) + ")"
-	}
-	if strings.HasPrefix(body, "putStrLn") {
-		arg := strings.TrimSpace(strings.TrimPrefix(body, "putStrLn"))
-		arg = strings.Trim(arg, "()")
-		if strings.HasPrefix(arg, "show ") {
-			arg = "str(" + strings.TrimSpace(strings.TrimPrefix(arg, "show ")) + ")"
-		}
-		return "print(" + convertExpr(arg) + ")"
-	}
-	return convertExpr(body)
-}
-
-var (
-	callParen   = regexp.MustCompile(`([A-Za-z_][A-Za-z0-9_]*)\s+\(([^()]*)\)`)
-	callBrack   = regexp.MustCompile(`([A-Za-z_][A-Za-z0-9_]*)\s+\[([^\]]*)\]`)
-	avgExpr     = regexp.MustCompile(`^fromIntegral\(sum\(([^()]*)\)\)\s*/\s*fromIntegral\(length\(([^()]*)\)\)$`)
-	mapFromList = regexp.MustCompile(`^Map\.fromList\(\[(.*)\]\)$`)
-)
-
-// fixFuncCalls converts Haskell-style function application to Mochi calls.
-func fixFuncCalls(expr string) string {
-	for {
-		changed := false
-		e := callParen.ReplaceAllString(expr, `$1($2)`)
-		if e != expr {
-			changed = true
-			expr = e
-		}
-		e = callBrack.ReplaceAllString(expr, `$1([$2])`)
-		if e != expr {
-			changed = true
-			expr = e
-		}
-		if !changed {
-			break
-		}
-	}
-	return expr
-}
-
-// trimOuterParens removes a single pair of wrapping parentheses if they
-// encompass the entire expression.
-func trimOuterParens(s string) string {
-	for {
-		if len(s) >= 2 && strings.HasPrefix(s, "(") && strings.HasSuffix(s, ")") {
-			depth := 0
-			ok := true
-			for i, ch := range s {
-				switch ch {
-				case '(':
-					depth++
-				case ')':
-					depth--
-					if depth == 0 && i != len(s)-1 {
-						ok = false
-						break
-					}
-				}
-			}
-			if ok && depth == 0 {
-				s = strings.TrimSpace(s[1 : len(s)-1])
-				continue
-			}
-		}
-		break
-	}
-	return s
-}
-
-// reorderVars moves variable declarations so that referenced variables are
-// defined before use. This handles simple dependencies between top-level vars.
-func reorderVars(items []Item) []Item {
-	nameIdx := map[string]int{}
-	for i, it := range items {
-		if it.Kind == "var" || it.Kind == "func" {
-			nameIdx[it.Name] = i
-		}
-	}
-	for i := 0; i < len(items); i++ {
-		it := items[i]
-		if it.Kind != "var" {
-			continue
-		}
-		for name, j := range nameIdx {
-			if j > i && strings.Contains(it.Body, name) {
-				items[i], items[j] = items[j], items[i]
-				nameIdx[it.Name], nameIdx[name] = j, i
-			}
-		}
-	}
-	return items
-}
-
-func convertExpr(expr string) string {
-	expr = strings.TrimSpace(expr)
-	expr = fixFuncCalls(expr)
-	expr = trimOuterParens(expr)
-	if strings.HasPrefix(expr, "\\") && strings.Contains(expr, "->") {
-		idx := strings.Index(expr, "->")
-		params := strings.Fields(strings.TrimSpace(expr[1:idx]))
-		body := strings.TrimSpace(expr[idx+2:])
-		var b strings.Builder
-		b.WriteString("fun(")
-		for i, p := range params {
-			if i > 0 {
-				b.WriteString(", ")
-			}
-			b.WriteString(p)
-			b.WriteString(": int")
-		}
-		b.WriteString("): int => ")
-		b.WriteString(convertExpr(body))
-		return b.String()
-	}
-	if strings.HasPrefix(expr, "trace ") {
-		rest := strings.TrimSpace(strings.TrimPrefix(expr, "trace "))
-		if i := strings.Index(rest, " "); i != -1 {
-			msg := trimOuterParens(rest[:i])
-			val := strings.TrimSpace(rest[i+1:])
-			return "__trace(" + msg + ", " + convertExpr(val) + ")"
-		}
-	}
-	if strings.HasPrefix(expr, "if ") && strings.Contains(expr, " then ") && strings.Contains(expr, " else ") {
-		condEnd := strings.Index(expr, " then ")
-		elseIdx := strings.LastIndex(expr, " else ")
-		if condEnd != -1 && elseIdx > condEnd {
-			cond := strings.TrimSpace(expr[len("if "):condEnd])
-			thenPart := strings.TrimSpace(expr[condEnd+len(" then ") : elseIdx])
-			elsePart := strings.TrimSpace(expr[elseIdx+len(" else "):])
-			return "if " + convertExpr(cond) + " then " + convertExpr(thenPart) + " else " + convertExpr(elsePart)
-		}
-	}
-	if strings.HasPrefix(expr, "fromEnum(") && strings.HasSuffix(expr, ")") {
-		inner := strings.TrimSuffix(strings.TrimPrefix(expr, "fromEnum("), ")")
-		return convertExpr(inner)
-	}
-	if strings.HasPrefix(expr, "fromEnum ") {
-		inner := strings.TrimSpace(strings.TrimPrefix(expr, "fromEnum "))
-		return convertExpr(inner)
-	}
-	if strings.HasPrefix(expr, "read ") && strings.Contains(expr, "::") {
-		parts := strings.SplitN(strings.TrimPrefix(expr, "read "), "::", 2)
-		val := strings.TrimSpace(parts[0])
-		typ := mapType(strings.TrimSpace(parts[1]))
-		switch typ {
-		case "int", "float", "string", "bool":
-			return typ + "(" + val + ")"
-		}
-	}
-	if m := mapFromList.FindStringSubmatch(expr); m != nil {
-		entries := strings.Split(m[1], "),")
-		var parts []string
-		for _, e := range entries {
-			e = strings.TrimSpace(e)
-			e = strings.TrimPrefix(e, "(")
-			e = strings.TrimSuffix(e, ")")
-			kv := strings.SplitN(e, ",", 2)
-			if len(kv) == 2 {
-				key := strings.TrimSpace(kv[0])
-				val := strings.TrimSpace(kv[1])
-				parts = append(parts, key+": "+val)
-			}
-		}
-		return "{" + strings.Join(parts, ", ") + "}"
-	}
-	if m := avgExpr.FindStringSubmatch(expr); m != nil {
-		if strings.TrimSpace(m[1]) == strings.TrimSpace(m[2]) {
-			return "avg(" + strings.TrimSpace(m[1]) + ")"
-		}
-	}
-	if strings.HasPrefix(expr, "[") && strings.HasSuffix(expr, "]") {
-		return expr
-	}
-	if strings.Contains(expr, "++") {
-		parts := strings.Split(expr, "++")
-		if len(parts) == 2 {
-			left := strings.TrimSpace(parts[0])
-			right := strings.TrimSpace(parts[1])
-			if strings.HasPrefix(right, "[") && strings.HasSuffix(right, "]") {
-				val := strings.TrimSuffix(strings.TrimPrefix(right, "["), "]")
-				return "append(" + left + ", " + strings.TrimSpace(val) + ")"
-			}
-		}
-	}
-	if strings.HasPrefix(expr, "div ") {
-		parts := strings.Fields(expr)
-		if len(parts) == 3 {
-			return parts[1] + " / " + parts[2]
-		}
-	}
-	if strings.HasPrefix(expr, "length(") && strings.HasSuffix(expr, ")") {
-		inner := strings.TrimSuffix(strings.TrimPrefix(expr, "length("), ")")
-		return "len(" + inner + ")"
-	}
-	if strings.Contains(expr, "`mod`") {
-		parts := strings.Split(expr, "`mod`")
-		if len(parts) == 2 {
-			return strings.TrimSpace(parts[0]) + " % " + strings.TrimSpace(parts[1])
-		}
-	}
-	expr = strings.ReplaceAll(expr, "True", "true")
-	expr = strings.ReplaceAll(expr, "False", "false")
-	if strings.ContainsAny(expr, "()") {
-		return expr
-	}
-	parts := strings.Fields(expr)
-	if len(parts) > 1 {
-		ops := map[string]bool{"+": true, "-": true, "*": true, "/": true, "%": true, "&&": true, "||": true, "==": true, "<": true, ">": true, "<=": true, ">=": true}
-		hasOp := false
-		for _, p := range parts[1:] {
-			if ops[p] {
-				hasOp = true
-				break
-			}
-		}
-		if !hasOp {
-			for i := 1; i < len(parts); i++ {
-				parts[i] = strings.TrimSuffix(parts[i], ",")
-			}
-			return parts[0] + "(" + strings.Join(parts[1:], ", ") + ")"
-		}
-	}
-	return expr
-}
-
-// convertItems converts the Haskell AST items to Mochi source code.
-func convertItems(items []Item) []byte {
-	items = reorderVars(items)
-	var out strings.Builder
-	needTrace := false
-	traceNames = make(map[string]bool)
-	for _, it := range items {
-		if strings.Contains(it.Body, "trace ") {
-			needTrace = true
-			traceNames[it.Name] = true
-		}
-	}
-	if needTrace {
-		for i, it := range items {
-			if it.Kind == "print" || it.Kind == "json" {
-				for name := range traceNames {
-					items[i].Body = replaceOutsideQuotes(it.Body, name, name+"()")
-				}
-			}
-		}
-	}
-	if needTrace {
-		out.WriteString("fun __trace(msg: string, v: bool): bool {\n")
-		out.WriteString("  print(msg)\n  return v\n}\n")
-	}
-	for _, it := range items {
-		switch it.Kind {
-		case "print":
-			out.WriteString("print(")
-			out.WriteString(convertExpr(it.Body))
-			out.WriteString(")\n")
-		case "json":
-			out.WriteString("json(")
-			out.WriteString(convertExpr(it.Body))
-			out.WriteString(")\n")
-		case "func":
-			out.WriteString("fun ")
-			out.WriteString(it.Name)
-			var paramTypes []string
-			var retType string
-			var convBody string
-			if it.Type != "" {
-				paramTypes, retType = parseSigTypes(it.Type)
-				convBody = convertExpr(it.Body)
-			} else {
-				paramTypes = make([]string, len(it.Params))
-				for i := range paramTypes {
-					paramTypes[i] = "int"
-				}
-				convBody = convertExpr(it.Body)
-				retType = "int"
-				if strings.HasPrefix(strings.TrimSpace(convBody), "fun(") {
-					retType = "fun(int): int"
-				}
-			}
-			out.WriteByte('(')
-			for i, p := range it.Params {
-				if i > 0 {
-					out.WriteString(", ")
-				}
-				out.WriteString(p)
-				if i < len(paramTypes) && paramTypes[i] != "" {
-					out.WriteString(": ")
-					out.WriteString(paramTypes[i])
-				}
-			}
-			out.WriteByte(')')
-			if retType != "" {
-				out.WriteString(": ")
-				out.WriteString(retType)
-			}
-			if it.Body == "" {
-				out.WriteString(" {}\n")
-			} else {
-				out.WriteString(" { return ")
-				out.WriteString(convBody)
-				out.WriteString(" }\n")
-			}
-		case "var":
-			if traceNames[it.Name] {
-				out.WriteString("fun ")
-				out.WriteString(it.Name)
-				out.WriteString("(): bool { return ")
-				out.WriteString(convertExpr(it.Body))
-				out.WriteString(" }\n")
-			} else {
-				out.WriteString("let ")
-				out.WriteString(it.Name)
-				if it.Type != "" {
-					if typ, ok := parseVarSig(it.Type); ok {
-						out.WriteString(": ")
-						out.WriteString(typ)
-					}
-				}
-				if it.Body != "" {
-					out.WriteString(" = ")
-					out.WriteString(convertExpr(it.Body))
-				}
-				out.WriteByte('\n')
-			}
-		case "for":
-			out.WriteString("for ")
-			out.WriteString(it.Name)
-			out.WriteString(" in ")
-			if it.Collection != "" {
-				out.WriteString(it.Collection)
-			} else {
-				out.WriteString(it.Start)
-				out.WriteString("..")
-				out.WriteString(it.End)
-			}
-			out.WriteString(" { ")
-			out.WriteString(convertExpr(it.Body))
-			out.WriteString(" }\n")
-		case "struct":
-			out.WriteString("type ")
-			out.WriteString(it.Name)
-			if len(it.Fields) == 0 {
-				out.WriteString(" {}\n")
-				break
-			}
-			out.WriteString(" {\n")
-			for _, f := range it.Fields {
-				out.WriteString("  ")
-				out.WriteString(f.Name)
-				if f.Type != "" {
-					out.WriteString(": ")
-					out.WriteString(mapType(f.Type))
-				}
-				out.WriteByte('\n')
-			}
-			out.WriteString("}\n")
-		}
-	}
-	if out.Len() == 0 {
-		return nil
-	}
-	result := []byte(out.String())
-	traceNames = nil
-	return result
-}
-
-func parseSigTypes(sig string) ([]string, string) {
-	sig = strings.ReplaceAll(sig, "\n", " ")
-	if i := strings.Index(sig, "::"); i != -1 {
-		sig = strings.TrimSpace(sig[i+2:])
-	}
-	if i := strings.Index(sig, "=>"); i != -1 {
-		sig = strings.TrimSpace(sig[i+2:])
-	}
-	if strings.HasPrefix(sig, "forall") {
-		if i := strings.Index(sig, "."); i != -1 {
-			sig = strings.TrimSpace(sig[i+1:])
-		}
-	}
-	parts := strings.Split(sig, "->")
-	for i, p := range parts {
-		parts[i] = strings.TrimSpace(p)
-	}
-	if len(parts) == 0 {
-		return nil, ""
-	}
-	ret := mapType(parts[len(parts)-1])
-	params := make([]string, 0, len(parts)-1)
-	for _, p := range parts[:len(parts)-1] {
-		params = append(params, mapType(p))
-	}
-	return params, ret
-}
-
-func mapType(t string) string {
-	t = strings.TrimSpace(t)
-	switch t {
-	case "", "()":
-		return ""
-	case "Int", "Integer":
-		return "int"
-	case "Float", "Double":
-		return "float"
-	case "String", "[Char]":
-		return "string"
-	case "Bool":
-		return "bool"
-	}
-	if strings.HasPrefix(t, "[") && strings.HasSuffix(t, "]") {
-		inner := mapType(strings.TrimSuffix(strings.TrimPrefix(t, "["), "]"))
-		if inner == "" {
-			inner = "any"
-		}
-		return "list<" + inner + ">"
-	}
-	return t
-}
-
-func parseVarSig(sig string) (string, bool) {
-	sig = strings.ReplaceAll(sig, "\n", " ")
-	if i := strings.Index(sig, "::"); i != -1 {
-		sig = strings.TrimSpace(sig[i+2:])
-	}
-	if i := strings.Index(sig, "=>"); i != -1 {
-		sig = strings.TrimSpace(sig[i+2:])
-	}
-	if strings.HasPrefix(sig, "forall") {
-		if i := strings.Index(sig, "."); i != -1 {
-			sig = strings.TrimSpace(sig[i+1:])
-		}
-	}
-	sig = strings.TrimSpace(sig)
-	if sig == "" {
-		return "", false
-	}
-	return mapType(sig), true
-}
-
-// ConvertSource converts parsed Haskell items into Mochi source code.
-func ConvertSource(items []Item, original string) (string, error) {
-	dup := make([]Item, len(items))
-	copy(dup, items)
-	src := convertItems(dup)
-	if src == nil {
-		return "", fmt.Errorf("no output")
-	}
-	var b strings.Builder
-	b.Write(meta.Header("//"))
-	b.WriteByte('\n')
-	b.WriteString("/*\n")
-	b.WriteString(original)
-	if !strings.HasSuffix(original, "\n") {
-		b.WriteByte('\n')
-	}
-	b.WriteString("*/\n")
-	b.Write(src)
-	return b.String(), nil
-}
-
-// Convert converts parsed Haskell items into a Mochi AST node.
-func Convert(items []Item, original string) (*ast.Node, error) {
-	dup := make([]Item, len(items))
-	copy(dup, items)
-	src, err := ConvertSource(dup, original)
-	if err != nil {
-		return nil, err
-	}
-	prog, err := parser.ParseString(src)
-	if err != nil {
-		return nil, err
-	}
-	return ast.FromProgram(prog), nil
-}
